Create test utils file
