+++ pyproject.toml
@@ -7,7 +7,7 @@ name = "FroggyTV"
 version = "0.0.1"
 
 [tool.pytest.ini_options]
-pythonpath = "src"
+pythonpath = "src tests/helpers"
 addopts = [
     "--import-mode=importlib",
 ]
